Move workload components to a separate package
