+++ csb/src/main/scala/edu/msstate/dasi/csb/workload/Neo4jWorkload.scala
@@ -1,4 +1,4 @@
-package edu.msstate.dasi.csb
+package edu.msstate.dasi.csb.workload
 
 import java.util.concurrent.TimeUnit
